Add photo feature.
On the info page for each team, there is now a "Photo" button next to the
team number. Pressing it opens the photo viewer, which contains a button
to take a new photo. The files are stored as JPEG images with the filename
TEAMNUMBER_TIMESTAMP.jpg. These images can be found on the emulated
external drive, in Android/data/team1100.pitscout2017/files/Pictures. To
remove images they currently need to be deleted manualy from the
filesystem. Like team information, the images can be viewed from any
competition the team is a part of.