Fix TeamList crash. Add 5 CIM option.
When I added the gear catcher option, there was a place where I apparently
did not increase the array size of stored data, so it was trying to put
the gear catcher info out of the array bounds. Also, 1247 has 5 drive
sims?!?!?! so I added that option, cause my life is terrible.